Bump v0.3.4+ dev

+++ evadb/catalog/services/index_catalog_service.py
@@ -35,9 +35,15 @@ class IndexCatalogService(BaseService):
         type: VectorStoreType,
         feat_column: ColumnCatalogEntry,
         function_signature: str,
+        index_def: str,
     ) -> IndexCatalogEntry:
         index_entry = IndexCatalog(
-            name, save_file_path, type, feat_column.row_id, function_signature
+            name,
+            save_file_path,
+            type,
+            feat_column.row_id,
+            function_signature,
+            index_def,
         )
         index_entry = index_entry.save(self.session)
         return index_entry.as_dataclass()
@@ -74,7 +80,11 @@ class IndexCatalogService(BaseService):
             index_metadata = index_obj.as_dataclass()
             # clean up the on disk data
             if os.path.exists(index_metadata.save_file_path):
-                os.remove(index_metadata.save_file_path)
+                if os.path.isfile(index_metadata.save_file_path):
+                    # For service-hosting-based vector database, we should not
+                    # touch their base directory. The only case that needs to
+                    # be taken care of is FAISS index local disk file.
+                    os.remove(index_metadata.save_file_path)
             index_obj.delete(self.session)
         except Exception:
             logger.exception("Delete index failed for name {}".format(name))
